tmux tab switching
action_switch_tab_cb is called to switch tabs (both Alt-<Num> and
Ctrl-PageUp/Down). Within this function, check if there is only one tab
open and, if there is, then replace the normal activity for these
keystrokes with tmux commands that try to achieve the same thing.

At present this is merely a quick hack, albeit one that feels great under
the fingers. So far I haven't come up with a way to reconcile this with
the gnome preferences ethos.